DPAD-996: allow setting icon size
