+++ src/experimental/shared/play-state/PauseButton.tsx
@@ -2,17 +2,9 @@ import React, { useContext } from 'react';
 import { PlayerContext } from 'jwplayer/players/shared/PlayerContext';
 import { IconWrapper } from 'experimental/shared/play-state/PlayStateWrapper';
 import PauseIcon from 'experimental/shared/icons/play/PauseIcon';
+import { PauseButtonProps } from 'experimental/types';
 
-export interface PauseButtonProps {
-	/** Optional, additional events that should fire whenever the pause button is pressed.
-	 *  The actual video player "pause" event will already be handled in the onClickPause function.
-	 *
-	 *  There is no need to call the player.pause() event in this callback */
-	onClickCallback?: () => void;
-	isAd?: boolean;
-}
-
-const PauseButton: React.FC<PauseButtonProps> = ({ onClickCallback, isAd }) => {
+const PauseButton: React.FC<PauseButtonProps> = ({ onClickCallback, isAd, iconSize }) => {
 	const { player } = useContext(PlayerContext);
 	const onClickPause = () => {
 		/* Tracking events should already be handled in a wrapping component.
@@ -26,7 +18,7 @@ const PauseButton: React.FC<PauseButtonProps> = ({ onClickCallback, isAd }) => {
 
 	return (
 		<IconWrapper onClick={onClickPause}>
-			<PauseIcon />
+			<PauseIcon height={iconSize || '16px'} width={iconSize || '16px'} />
 		</IconWrapper>
 	);
 };
