+++ src/experimental/shared/play-state/PlayButton.tsx
@@ -2,17 +2,9 @@ import React, { useContext } from 'react';
 import { PlayerContext } from 'jwplayer/players/shared/PlayerContext';
 import { IconWrapper } from 'experimental/shared/play-state/PlayStateWrapper';
 import PlayIcon from 'experimental/shared/icons/play/PlayIcon';
+import { PlayButtonProps } from 'experimental/types';
 
-export interface PlayButtonProps {
-	/** Optional, additional events that should fire whenever the Play button is pressed.
-	 *  The actual video player "play" event will already be handled in the onClickPlay function.
-	 *
-	 *  There is no need to call the player.play() event in this callback */
-	onClickCallback?: () => void;
-	isAd?: boolean;
-}
-
-const PlayButton: React.FC<PlayButtonProps> = ({ onClickCallback, isAd }) => {
+const PlayButton: React.FC<PlayButtonProps> = ({ onClickCallback, isAd, iconSize }) => {
 	const { player } = useContext(PlayerContext);
 	const onClickPlay = () => {
 		/* Tracking events should already be handled in a wrapping component.
@@ -26,7 +18,7 @@ const PlayButton: React.FC<PlayButtonProps> = ({ onClickCallback, isAd }) => {
 
 	return (
 		<IconWrapper onClick={onClickPlay}>
-			<PlayIcon />
+			<PlayIcon height={iconSize || '16px'} width={iconSize || '16px'} />
 		</IconWrapper>
 	);
 };
