+++ src/experimental/shared/volume-state/VolumeStateWrapper.tsx
@@ -49,6 +49,7 @@ const VolumeStateWrapper: React.FC<VolumeStateWrapperProps> = ({
 	hasSlider = false,
 	hasLabel = false,
 	callback,
+	iconSize,
 }) => {
 	const { player } = useContext(PlayerContext);
 	const mute = useMute();
@@ -66,7 +67,11 @@ const VolumeStateWrapper: React.FC<VolumeStateWrapperProps> = ({
 		<Wrapper onClick={onClick} color={iconColor}>
 			<SoundButtonWrapper onMouseEnter={() => setHover(true)} onMouseLeave={() => setHover(false)}>
 				{hasSlider && <VolumeSlider color={sliderColor} hover={hover} />}
-				{mute ? <StyledSoundOffIcon fill={iconColor} /> : <StyledSoundIcon fill={iconColor} />}
+				{mute ? (
+					<StyledSoundOffIcon fill={iconColor} width={iconSize || '16px'} height={iconSize || '16px'} />
+				) : (
+					<StyledSoundIcon fill={iconColor} width={iconSize || '16px'} height={iconSize || '16px'} />
+				)}
 			</SoundButtonWrapper>
 			{mute && hasLabel && <Label>Play Sound</Label>}
 		</Wrapper>
